Refactor Error Handling for VaDotGovService (#11479)
`VaDotGovService` provides an interface to 2 `va.gov` API services

1. Vet360 - validates an address, returning standardized postal addresses and a geolocation
2. Facilities API - returns data related to VA facilities and, if lat/long coordinates are provided, distances to each facility

`VaDotGovAddressValidator` implements logic for finding and storing a veteran's closest regional office as well as distances to VA hearing facilities. 

Previously, application flow was controlled through raising and handling a series of exceptions. In addition to being [slower](https://www.honeybadger.io/blog/benchmarking-exceptions-in-ruby-yep-theyre-slow/), this approach made the code hard to follow and hard to add additional steps to the logic.

This PR refactors `VaDotGovService`/`VaDotGovAddressValidator` in the following ways:
- Moves all response handling to helpers classes `FacilitiesResponse` and `AddressValidationResponse`
- All public methods on `VaDotGovService` return a hash with an `error` property defined
  - `get_distance` - `{ error: Caseflow::Error, facilities: [] }`
  - `validate_address` - `{ error: Caseflow::Error, valid_address: {} }`
- Adds `ErrorHandler` to `VaDotGovAddressValidator` that handles creating admin actions for appeals where a closest regional office can't be found
- Adds a `Validations` module that checks for errors in results for `va.gov`