Stop sending Slack job alerts from UAT (#11703)
Extends the work done in #11628. This PR will make it so that jobs that run in UAT do not send any notifications to Slack. Below is a snapshot of the `#appeals-app-alerts` channel showing notifications from jobs running in UAT and Prod. The second and third notifications need developer attention, but they are sandwiched by notifications about jobs running in UAT that do not need our attention. This PR aims to reduce the amount of noise we send to the `#appeals-app-alerts` channel.

![image](https://user-images.githubusercontent.com/32683958/62706450-8ef6b080-b9bd-11e9-87b1-2ea64cfa4e3d.png)