Refactor SlackService
**Why**:
- To make the tests less brittle by using dependency injection. If we
end up using a different HTTP library, such as Faraday, we won't have
to change the tests
- To eliminate the use of `allow_any_instance_of`, which is discouraged
by the RSpec maintainers
- To use named keywords so that's it's clear what each argument
represents
- To fix the FeatureEnvy code smell by making the msg, title, and
channel arguments part of the initialization as opposed to arguments
for the `send_notification` method.

This PR also improves test coverage for SlackService, which revealed
a bug when the channel argument did not start with the # character.

Further improvements could be to rename the class from the generic
`SlackService` to something more precise such as `SlackNotification`
and then rename the `send_notification` method to `call`.