Do not include canceled claims in taken EP modifiers (#11587)
Cancelled claims should be released after they are cancelled, like cleared claims. We temporarily did this because many canceled claims were being affected by an external issue where the modifiers weren't being released (which also can affect cleared claims but was less common).

I want to fix this to the intended state (especially since there's no comment explaining the above).  Also, we are now trying to report the Veterans who have this issue to BGS so that they can fix them, but it's difficult to know which Veterans are affected unless they experience the error on a claim that wasn't closed recently.