use zero-padded day of month to match moment.js (#11958)
# Description

Was seeing this error in CircleCI:

```
Failure/Error: expect(page).to have_content("Read #{message.read_at.friendly_full_format}")
```

The issue was around the date time in the test looking differently than the one in the UI due to a discrepancy in how the values were formatted:

```
Tue Sep  3 2019 at 11:35
Tue Sep 03 2019 at 11:35
```

One is 0-padded, the other is not. I updated the format the test was checking for to match what was being shown.