Exclude TimedHoldTasks from bulk assignment (#12038)
Resolves bug exposed by [this sentry alert](https://sentry.ds.va.gov/department-of-veterans-affairs/caseflow/issues/6516/) (and [related slack conversation](https://dsva.slack.com/archives/C4JECDLSE/p1568136167020400)).

The bug here is that the application attempts to create child tasks for `TimedHoldTask`s assigned to the `HearingsManagement` organization. That attempt will fail because `TimedHoldTask`s [should not have any child tasks](https://caseflow-looker.va.gov/sql/dns58vwdwyszry) (ticketed in #12037) and, more specifically, the request from the front-end does not pass a field required by the `TimedHoldTask` (`days_on_hold`) so validation thankfully fails to create the invalid task.

This PR prevents this issue by excluding `TimedHoldTask`s from being returned to the front-end as an option in the dropdown menu.