Begin unifying GenericTask and Task models. (#12042)
Bumps #11607 

This PR hollows out the `GenericTask` class, moving all functionality into the `Task` model class. It also refactors all of the `GenericTask` subclasses to instead be subclasses of `Task`. 

Finally, it modifies one test in `task_spec.rb` to account for the fact that the updated `Task.create_from_params` has been updated to use the method formerly used in `GenericTask`. This test change is low stakes -- nothing in our application code was using the original `create_from_params`, only the specs in `task_spec.rb`.

Work left to do if this survives in production for a few days with no incidents:

- [ ] Update database to change all rows in `tasks` table with a `type` of `GenericTask` to have type `Task` isntead
- [ ] Eliminate `GenericTask` class entirely and update spec code to reflect as much, including modifying factories and merging `generic_task_spec.rb` into `task_spec.rb`