Optimize most recently held hearing controller action (#12090)
Related to #12002 

### Description

The `/appeal/:id/hearings` is overall one of the most time consuming calls in Caseflow. It's a fairly simple action, which probably means it's called a lot:

<img width="1118" alt="Screen Shot 2019-09-13 at 4 18 13 PM" src="https://user-images.githubusercontent.com/1298274/64892400-509a8400-d642-11e9-959d-013a778718a6.png">

The actual code is just finding a list of associated hearings for a specific appeal, and getting the most recent one that was held. For legacy appeals, the route triggers a bunch of VACOLS calls:

<img width="717" alt="Screen Shot 2019-09-13 at 4 21 57 PM" src="https://user-images.githubusercontent.com/1298274/64892532-a2dba500-d642-11e9-916a-eeedacda1ab0.png">

4/5 of those VACOLS calls are triggered from getting the appeal. Using `kcachegrind` and `ruby-prof`, it was really easy to see this:

<img width="864" alt="Screen Shot 2019-09-13 at 4 25 20 PM" src="https://user-images.githubusercontent.com/1298274/64892687-1bdafc80-d643-11e9-9f5c-7678aa01aa2c.png">

Roughly 80% of the time spent in the controller action is spend in `AppealsController::appeal`, which is responsible for getting a legacy appeal from VACOLS.

The controller action isn't actually using the vast majority of the data we're getting back from VACOLS other than the hearings data. I changed the route to make an assumption that if there are hearings in VACOLS, there probably is a case in VACOLS as well. This lets us skip 4/5 of the VACOLS calls in most cases. The only time these VACOLS calls are still made is if the appeal doesn't exist yet in the Caseflow database.

After these changes, the vast majority of the time is spent trying to get the hearings, and the `wall_time` is cut roughly in half:

<img width="864" alt="Screen Shot 2019-09-13 at 4 30 54 PM" src="https://user-images.githubusercontent.com/1298274/64893006-f39fcd80-d643-11e9-8fa5-3a281892f469.png">

#### Changes

  * Change the name of the controller method to match more what it is doing
  * Create a serializer for the hearings hash
  * Only get VACOLS Case data if no VACOLS hearings are found