Implement naive caching for queue pagination API (#12131)
Connects #11310. Adds naive caching based only on the URL of the endpoint we are making the task page request to. Later on, we can add the ability to bypass the API request when there is only a single page of results (thus the front-end already has access to all of the rows that the back-end could possibly return so we can do all of the filtering on the front-end without needing to involve the back-end).

In the screen capture below, notice that the second time we make a request for hearing clarification tasks we do not make an API request.

![naive_caching](https://user-images.githubusercontent.com/32683958/65074122-29092b80-d962-11e9-849e-fcfebf44ddf1.gif)