Add Hearings Page: Screen Reader Modal Navigation Fix (#12196)
Resolves #12121 

This fix selects the first focusable element and the last selectable element in the modal and cycles the user through each option when they press `Tab`.

In addition, an extra focus target has been added to `compnentWillUnmount`. This will place the focus back on to the button that triggered the modal to open instead of navigating to `Digital Services at VA` at the bottom of the page.