instead of sleep, cache geomatch API responses (#12366)
Our `rake db:seed` task was very slow because we had two `sleep 1` calls in a loop, as part of the `FetchHearingLocationsForVeteransJob` process.

This PR adds some logging to the seed process so it's easier to spot which steps are happening in the Rails logs, and eliminates the `sleep` calls. Instead identical API calls to the geomatching service are cached for 24 hours, which avoids the need I think the `sleep` was filling (throttling our use of that va.gov API). My assumption with the caching logic is that geomatching is a relatively fixed thing (stations don't physically move often).

The speed up is pretty dramatic:
```
before: 5m52.478s
after: 2m11.036s
```