Fix remanded definition for request issue (#12426)
### Description
This fixes the definition of "remanded?" for request issue.

If the request issue is correcting a decision issue, instead of checking if the contested decision issue is remanded, check if the decision issue's decision review is a remand supplemental claim.  This is because the correction could be on a decision issue that is "granted" or "denied", but since it's a correction it's not being contested on a new claim.  It's being contested on the same decision review, and it should still go onto a "remand" type EP code.

### Documentation Updates
- [x] Topline File Descriptions Added or Updated as Needed