Save end product claim type code updates (#12427)
### Description
This table is to save end product claim_type_code updates when they are changed outside of Caseflow manually (in Share).  These are causing us errors in processing remand supplemental claims, and are potentially causing eligibility check errors and other issues as well.

The initial purpose of this table is to be able to do some investigation and reporting on the problem.  We do not yet have solutions for how to resolve the problems, but it will hopefully useful for implementing solutions. If not, it may be a temporary table.

### Documentation Updates
- [x] Topline File Descriptions Added or Updated as Needed

*Schema Changes Only*

* [x] Timestamps (created_at, updated_at) for new tables
* [x] Column comments updated
* [ ] Query profiling performed (eyeball Rails log, check bullet and fasterer output)
* [ ] Appropriate indexes added (especially for foreign keys, polymorphic columns, and unique constraints)
* [x] #appeals-schema notified with summary and link to this PR