Create QueueTabs for individual queues (#12472)
Connects #11698. 1 of what will likely be 3 PRs associated with that ticket.

Individual and organizational queues display different columns for assigned, on hold, and completed queues. Instead of adding conditional logic to display different columns in the three `QueueTab` subclasses associated with those tabs, this PR creates new `QueueTab` subclasses for the individual queues.