Optimize Hearing Days Index a bit (#12084)
### Description

I'm trying some ways to learn how to profile performance for a Rails controller. Using `ruby-prof` and `kcachegrind` I was able to find a call that probably could be a little faster. Ultimately it's a very minor improvement! Could eventually work on ways to introduce this into the development process.

#### Changes

  - Prefetch judge for calls that use `HearingDay::to_hash`
  - Create `HearingDaySerializer` to use for `to_hash` instead of ActiveRecord serializer

#### Before

I noticed that `HearingDay::to_hash` was pretty high in the list of expensive calls after profiling `/hearings/hearing_day.json`:

![Screen Shot 2019-09-12 at 5 29 42 PM](https://user-images.githubusercontent.com/1298274/64822411-eec71500-d582-11e9-8226-c8025046c4a9.png)

Clicking through, I found that a significant portion of that was in `HearingDay::judge_first_name`. It seemed odd because `HearingDay::judge_last_name` was much quicker:

![Screen Shot 2019-09-12 at 5 31 31 PM](https://user-images.githubusercontent.com/1298274/64822521-2e8dfc80-d583-11e9-8765-c14ce14298c8.png)

The slowness appears to be because `judge_first_name` is triggering a call to get the judge user from the database for the first time.

#### After

After these changes, `HearingDay::to_hash` drops way down in the list:

![Screen Shot 2019-09-12 at 5 35 27 PM](https://user-images.githubusercontent.com/1298274/64822822-dc011000-d583-11e9-988d-9557d46af14e.png)

The attribute serialize method for `judge_first_name` and `judge_last_name` are pretty quick afterwards, and roughly the same, which we would expect from prefetching:

![Screen Shot 2019-09-12 at 5 38 03 PM](https://user-images.githubusercontent.com/1298274/64822927-24b8c900-d584-11e9-8582-09fe2c0a1c13.png)

Also the `wall_time per call` column changes from `15 881` for `ActiveRecord/Delegation::each` to `1 850`! I think the units for those measurements is in microseconds though so it might be a small portion of the actual time spent in the controller. Still cool to see the changes!