MTV | Add option for partial grant and select issues (#12454)
Connects #12318

### Description
Updated various bits related to judge "Address Motion to Vacate" view to support "partial vacatur" as a disposition. This also enables selection of issues in case of "partial" disposition.

This PR also adds backend functionality for allowing "partial" disposition so that things don't break. Backend handling of the selected issues will be handled in a future PR.

### Acceptance Criteria
- [ ] Code compiles correctly
- [ ] Address Motion to Vacate supports `partial` disposition
- [ ] Address Motion to Vacate displays and mandates selection of one or more issues when `partial` disposition is selected