Create DecisionDraftingAttorney rows when adding users to JudgeTeams (#12493)
Connects #12481. Whenever a user is added to a `JudgeTeam`, create a corresponding `DecisionDraftingAttorney` instance.

For now, we are assuming all users added to `JudgeTeam`s will be attorneys. This may change in the future when we roll out more functionality for DVCs. When we do that, we may need to modify this function.