Remove json_hearing_day and json_hearing_days (#12727)
Resolves #12622

### Description

I decided to do this in such a way where the structure of the data that the API returned was intact.

This led to some issues because `HearingDay#to_hash` was being used in `assign_judges_to_hearing_days.rb`. The change required updating `assign_judges_to_hearing_days.rb` to no longer depend on the behavior of `HearingDay#to_hash`. The file was tested pretty extensively, so I'm hoping the unit tests catch any regressions.

#### Changes

  * Remove `HearingDayController#json_hearing_days` and `HearingDayController#json_hearing_day`
  * Remove `HearingDay#create_hearing_day`, which chained together `HearingDay#create` and `HearingDay#to_hash`. IMO, ideally, `HearingDay#to_hash` should primarily be used as a representation of `HearingDay` to the frontend. We should avoid using a hash to represent a `HearingDay` in the backend portions of the code.
  * Update `HearingDaySerializer` to serialize `regional_office` and `room` and keep the structure of the data intact.
  * Refactor `assign_judges_to_hearing_days.rb` to use HearingDay model instead of hash
  * Remove `app/services/hearing_schedule/generate_co_hearing_days_schedule.rb`. This was an unused piece of code that called `HearingDay#create_hearing_day`
  * Update tests to expect the `HearingDay` model instead of a hash