Use request issues update user for establishment and cancellations (#12758)
### Description
When a user edits the issues on a claim, they can add and remove issues.  This can subsequently result in adding or removing end products.  These establishments or cancellations should be associated to the user who performed the update.

Currently, the user is manually being set to the system user. This PR sets the user to the RequestIssuesUpdate user, if a RequestIssuesUpdate is the source of the DecisionReviewProcess job.

Please note: This is the first (small change), of what may be many changes with respect to stopping from using the system user. This specific request is for EP cancellations, so the user has been updated for RequestIssuesUpdates only in this PR. More time and investigation is necessary before making additional changes, such as automatically generated decision reviews, EPs established from intakes, or whatever else BGS brings to our attention.