Include error message in ACD docket number error (#12774)
The code comment two lines below this error explains why we are throwing this error. This PR includes that comment in the error itself to be more resilient to code changes in the future that may cause the code comment to drift away from the place where we are throwing the error.