implement sort fix with role & tabIndex (#12898)
Partially addresses #12126

### Description
These changes allow both VoiceOver and JAWS to identify & interact with the sort functionality utilizing the keyboard.  

A secondary aspect to the original ticket was to implement a fix for keyboard functionality with the filter menu.  Voiceover functionality with this was pre-existing however a fix for integration with JAWS has not been identified or implemented.