Include vet file numbers in async job labeling for Inbox (#13000)
Connects #12926 

### Description
As part of AMO's feedback on the async job messaging demo, in Inbox prose, labels for jobs will contain veteran file numbers.

Because there are several existing places (e.g. the `/jobs` index and specific job pages) where the job ID is already used, I chose to err on the side of extra information in the Inbox for the sake of consistency.

### User Facing Changes
Sample screenshot of an Inbox view with the new labels:
<img width="1089" alt="Screen Shot 2019-12-17 at 13 18 58" src="https://user-images.githubusercontent.com/282869/71023164-2ab4cd80-20d0-11ea-8afc-b1d301f74fb1.png">
No labeling outside of the Inbox has been changed.