ETL::OrganizationsUser sync by unique cols index (#13020)
Discovered while syncing in production, organizations_users rows are deleted and re-created for the same org/user pair. That will change the primary key id value (since it's a new record). The ETL table will get out of sync if we rely on 1:1 match for `id` columns.

Instead we use the unique index columns to match. The `id` can therefore change for the "same" record but that's ok because this table is a m2m joining table and the `id` is ephemeral.