Decouple UpcomingHearingDays from AssignHearingTabs (#13107)
Prerequisite to #11757

Description
Moves UpcomingHearingsTable to a separate component, which doesn't use AssignHearingTable
Fix bug in PowerOfAttorneyDetail where the component thought it was constantly being loaded
Pass the current row to the filterValueTransform. This acts as context, and can optionally be handled
I'm making this change because it is not necessary to paginate UpcomingHearingsTable. The data that powers the UpcomingHearingsTable is bounded by the number of hearings for a given day (usually fewer than 20). UpcomingHearingsTable leverages AssignHearingsTable, and the plan right now is to add pagination to just AssignHearingsTable. It doesn't make sense to make AssignHearingsTable flexible enough to handle both cases...and I think it's probably just better to separate them at this point.

* pass row as second argument into filterValueTransform

* fix error with power of attorney never displaying

* refactor UpcomingHearingsTable to separate component. decouple from AssignHearingsTable

* address lint warnings

* address lint issues; add styling to tables

* add react import

* ordering imports

* reordering imports

* remove some unnecessary column fields

Co-authored-by: Peter Karman <pkarman@users.noreply.github.com>