Add docket_number and stream_type columns to Appeals table (#13241)
Connects #13211 

### Description
This is the first of multiple PRs for adding (MTV-required) support for multiple appeal streams with the same docket number. Previously, each appeal had a unique docket number and a type of `Original`. Going forward, the MTV flow will create a new appeal with the same docket number as the original, but with type `Vacate`.

This PR just adds the two DB columns and updates the `Appeal` model to use the values of those columns when not nil, or fall back on the existing behavior of hard-coded ephemeral values.

Future work:
- Update existing appeal-creation code to persist values to DB
- Backfill these two columns for all existing appeals
- After backfilling, remove the logic that supplies default values when DB values are nil

Note:
ActiveRecord is not happy about a DB column named `type`, so I named it `stream_type` (`docket_type` is used for a different purpose). We may want to consider changing the Appeal model's method name anyway, for the sake of consistency and because "type" is quite generic.

### Acceptance Criteria
- [ ] Code compiles correctly
- [ ] `appeals` table contains new columns
- [ ] `Appeal#docket_number` and `#type` use existing behavior when new columns are nil
- [ ] `Appeal#docket_number` and `#type` use values in new columns when they're not nil

### Testing Plan
1. Tested in Rails console:
```
> a = Appeal.first
> a.docket_number
=> "191208-1"
> a.type
=> "Original"
> a = Appeal.new(docket_number: "foo", stream_type: "Vacate")  # or use psql to insert
> a.docket_number
=> "foo"
> a.type
=> "Vacate"
```

### Database Changes
*Only for Schema Changes*

* [x] Column comments updated
* [x] #appeals-schema notified with summary and link to this PR