Persist stream docket number and type to new database columns (#13258)
Connects #13211 

### Description
Second in a series of PRs (first: #13241) to support appeal streams by storing the docket number and type in the DB.

This PR adds a `before_save` hook to Appeal which persists accurate values for docket number and type to the DB. After this PR, the DB fields will be populated for new and touched appeals, including MTV appeal when they land.

N.B.: with docket number, there's a tension in the fact that a non-nil stored value overrides the computed value, which in turn overrides a nil stored value — possible if the Appeal is new or only recently got its `id` or `receipt_date` — and this has some subtle order-of-operations implications. 

I've added a test for edge cases, even though the happy path for appeal creation is pretty well-behaved. Other tests are somewhat less well-behaved.

### Acceptance Criteria
- [ ] Code compiles correctly

### Testing Plan
1. Added a model test, to stress-test the order-of-operations subtleties mentioned above.