Fix error caused by undefined appeals object in SelectRemandReasonsView (#13337)
Resolves #13216 

### Description
We were attempting to access attributes on a non-existent appeals object within the mapStateToProps function which is called anytime the Redux store updates.  Specifically, when the user cancelled the review process it triggered the `cancelFlow` function and subsequent `CHECKOUT_STAGED_APPEAL` action in QueueFlowPage which removed the appeal object from `state.queue.stagedChanges`.  

The code needed to be adjusted to check for the existence of the object prior to interacting with it.