removed '.json' extension from imports (#13357)
### Description
This is a follow-up to the PR (#13347) that allows importing of JSON files w/o using the extension (this matches Webpack defaults, and makes JSON import syntax match that for JS/JSX imports). That PR also included an update to eslint config to allow for those, but in doing so it *prohibited* the '.json' extensions, which will trigger linting warnings on any modified files that include them. 

This PR does a global removal of the '.json' extension from all imports in files that *don't already have other linting errors*. This avoids triggering those same linting warnings with a PR designed to fix them. :P