Upstream Error | Do not fetch contentions for canceled EP to prevent FileNumberNotFoundForClaimID error (#13395)
connects #11787

### Description
Some EPs get canceled before contentions have been created. In this case, when we try to sync them, fetching the contentions fails, returning the FileNumberNotFoundForClaimId error.

This PR does not try to fetch the contentions or sync the source if the EP has been cancelled, instead just updating the status and closing the request issues.