Uncaught js XHR error (#13458)
connects #13077 

sentry error https://sentry.ds.va.gov/department-of-veterans-affairs/caseflow/issues/7774

is an uncaught non-200 XHR response. Instead of letting it get to Raven as a generic exception, log it to the console (at least) with a more helpful error.

This happens when a hearing has been deleted in VACOLS and Caseflow is out of sync. That should be caught by an actionable data integrity checker.

Example console output:

<img width="466" alt="Screen Shot 2020-02-19 at 3 29 15 PM" src="https://user-images.githubusercontent.com/1205061/74878288-a2031980-532c-11ea-891f-1a779736ae84.png">