Update rc-collapse dependency (#13502)
connects #12143

### Description
A button was given the "tab" role. This was happening in a dependency called rc-collapse.  When checking out their module, the latest version has fixed this, so this PR just upgrades to the latest version.

In the latest version, if there can be multiple panels active at the same time, then it becomes a tab.

This component is also used in HearingWorksheetDocs.  Looking at the code, I can't see a problem with it being coded as a button instead of a tab there, but wanted to mention it as a side effect.