Use the presence of a decision doc id to determine if a case has already gone to an attorney (#13530)
Resolves [these](https://dsva.slack.com/archives/CHX8FMP28/p1581619301317600) batteam [issues](https://dsva.slack.com/archives/CHX8FMP28/p1577821139163100) 

### Description
It looks like looking for an associated attorney case review is not enough to determine if a legacy case decision has already been drafted. Instead, we'll look at the decass record for the case to see if there is a decision document. 

### Notes
In testing I saw no ill effects when dispatching a case as an AVLJ when there was a missing attorney case review

### Testing plan
1. Sign in as BVASCASPER1 and draft a decision for a legacy appeal
2. At the end, send this case to an acting judge AVLJ - Acting judge
3. Sign in as said judge
4. Delete the associated case review
   ```ruby
   vacols_id = <vacols_id>
   LegacyAppeal.find_by(vacols_id: vacols_id).attorney_case_review.destroy!
   ```
5. Ensure the case shows up in the user's queue as a Judge review task
6. Ensure the user has the option to send this case to dispatch
7. Ensure completing judge checkout completes successfully