Explicitly prefer BGS POA on cached appeals attributes (#13554)
connects #13384 

### Description
We prefer displaying BGS POA name, so explicitly cache that. Otherwise, the `PowerOfAttorney` model will trigger loads of VACOLS `case_record` and `representative` that we will not use.

If BGS POA is not found, fall back to VACOLS.

In my manual tests in production, this greatly reduces calls to VACOLS, and instead shifts POA info to the BGS service (which is now cached for 30 days in Redis).