Intake | Fix bug when correcting an issue on a DTA supplemental claim. (#13558)
#### Description
When a request issue gets decided with a DTA disposition, it creates a remand supplemental claim automatically contesting the DTA disposition in order for that issue to be sent to another team.  If someone goes to the original review to make a correction on the DTA disposition, like if it was supposed to be granted or denied, then we want to remove the request issue contesting the DTA disposition. This allows for creating the new request issue contesting the DTA decision (but this time it's the issue to correct it to "granted").

Another case is when a DTA issue on a remand supplemental claim gets processed as usual and gets a decision, such as granted or denied.  If that issue gets a correction, even though it's a granted issue, it's still considered a remand issue because it's on a remand_supplemental_claim, at least for the purposes of EP creation.  In this cases, when the correction is being processed, when it's looking for other remanded issues contesting the same decision, it's finding itself, then it tries to remove itself (thinking it was an issue in the above scenario).

This PR prevents this behavior, so that the new correction_request_issue isn't then promptly removed.

This was causing a bug in Batteam, due to not being able to find an end product to cancel when trying to remove the new issue.
https://dsva.slack.com/archives/CHX8FMP28/p1582899519074600