Intake | Use intake user to set station of jurisdiction (#13650)
Connects #13247

### Description
When establishing an EP, this uses the intake user's station as the station of jurisdiction. Previously, we were setting the station of jurisdiction to 499 for the National Work Queue.  These were getting brokered to the station of the user submitting the establishment request to VBMS, which was often the system user. These were going to the Hines station instead of a regional office.

In the case of remand supplemental claims, this value is overridden to station 397 in the SupplementalClaim model.