Intake | Use last_synced_at when an EP's last action date is missing (#13651)
### Description
When we sync decision issues for non-rating issues, we depend on the End Product's last action date to approximate when the issue was decided.

I've discovered that sometimes even for cleared EPs that the last_action_date data is missing from BGS.  This is preventing decision issue syncing.

Use of the last_action_date is already an approximation of the issues decision date, since we have no other date information for when the decision was created. Dispositions do not have a date, and there is not a new rating created for non-rating issues.

We sync active EPs daily until they get a status of "CLR" or "CAN".  So the "last_synced_at" is something that we can use to approximate the last_action_date when it's missing.