Add diagnostic code description (#13741)
### Description
Adds a description for diagnostic code 7712, in response to this [batteam request](https://dsva.slack.com/archives/CHX8FMP28/p1584554603365600).

The values for staff_description and status_description come from the dgnstcTn and dgnstcTxt from data in the rating profile from BGS.  I'm not sure if there is another preferred source for this data.  I checked one other code that was available in the same rating profile and these fields were not a perfect match.